# src/api.yml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Usage api
  description: Usage api service APIs
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /metering
paths:
...
  /v1/usageevent:
    post:
      description: Super user access to usage event collection. You can fetch usage event by providing id or statusId or combination of usage event id and accountId.
      operationId: getUsageEvent
      x-eov-operation-handler: usage-events
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                accountOrPrefix:
                  type: string
                statusId:
                  type: string
                eventId:
                  type: string
      responses:
        '200':
          description: Usage Event object
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsageEvent'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      required:
        - errorCode
        - message
      properties:
        errorCode:
          oneOf:
            - type: string
            - type: number
        message:
          oneOf:
            - type: string
            - type: number
        details:
          type: object
...
    UsageEvent:
      required:
        - statusId
        - usage
        - enrichment
      properties:
        id:
          type: string
        statusId:
          type: string
        accountOrPrefix:
          type: string
          nullable: true
        usage:
          type: object
          properties:
            eventId:
              type: string
          additionalProperties: true
          required:
            - eventId
        enrichment:
          $ref: '#/components/schemas/Enrichment'
    Enrichment:
      properties:
        isAmendment:
          type: boolean
        productId:
          type: string
        editionId:
          type: string
        isTestSubscription:
          type: boolean
        isIBMProduct:
          type: boolean
      additionalProperties: true
...